+++ tests/files/src/0596_no_split.php
@@ -1,0 +1,4 @@
+<?php
+var_export(split(',', 'foo,bar'));
+var_export(spliti(',', 'foo,bar'));
+var_export(ereg('(0-9){4}', '1234'));